fixed unit by advices from liyang
